fix breaking texture change
When changing the name of the sampler2D variable in the shader code, it also needed to be changed here as well, so the Uniform object knows the name of the variable to look for during the locateUniforms() call. The previous version appeared to work just fine with a single texture (which defaults to texture unit #0), but would break when using multiple textures.